deployment/requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==41.0.1
fonttools==4.39.4
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.3
paramiko==3.2.0
pycparser==2.21
PyNaCl==1.5.0
pysftp==0.2.9
python-dotenv==1.0.0
